Fix CKAN roles resolution order
It did not consider the possibility that we give several CKAN roles to a given
user. It picked the first identified role, and the order was, in fact,
random-ish.
If a user was given, for instance, CKAN_EDITOR and CKAN_ADMIN, the retained role
might have been `editor`
This fix improved the matching and will give the highest role given to the user.